Moved button onclick functions to GMapsActivity so that they'll work.
