+++ app/src/main/java/com/example/lightway/GMapsActivity.java
@@ -1,5 +1,6 @@
 package com.example.lightway;
 
+import android.content.Intent;
 import android.location.Location;
 import android.support.v4.app.FragmentActivity;
 import android.os.Bundle;
@@ -105,7 +106,7 @@ public class GMapsActivity extends FragmentActivity implements OnMapReadyCallbac
 
         // Build the map.
         SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
-                .findFragmentById(R.id.Gmap);
+                .findFragmentById(R.id.gMap);
         mapFragment.getMapAsync(this);
 
     }
@@ -179,7 +180,7 @@ public class GMapsActivity extends FragmentActivity implements OnMapReadyCallbac
             public View getInfoContents(Marker marker) {
                 // Inflate the layouts for the info window, title and snippet.
                 View infoWindow = getLayoutInflater().inflate(R.layout.custom_info_contents,
-                        (FrameLayout) findViewById(R.id.Gmap), false);
+                        (FrameLayout) findViewById(R.id.gMap), false);
 
                 TextView title = infoWindow.findViewById(R.id.title);
                 title.setText(marker.getTitle());
@@ -410,5 +411,15 @@ public class GMapsActivity extends FragmentActivity implements OnMapReadyCallbac
         } catch (SecurityException e)  {
             Log.e("Exception: %s", e.getMessage());
         }
+    }
+
+    public void airStationsAPIActivity(View view) {
+        Intent intent = new Intent(this, AirStationsAPIActivity.class);
+        startActivity(intent);
+    }
+
+    public void parkingAPIActivity(View view) {
+        Intent intent = new Intent(this, ParkingAPIActivity.class);
+        startActivity(intent);
     }
 }
