+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     tools:context="com.example.lightway.MainActivity">
 
     <fragment
-        android:id="@+id/Gmap"
+        android:id="@+id/gMap"
         android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="match_parent"
         android:layout_height="0dp"
